Invalidate rather than update cache
Previously on touch or write events we called `_ls` to update the local
cache.  When we called touch repeatedly this caused a slowdown unless
we used the `no_refresh` context manager.

Now we just invalidate the cache on write events so that the next read
event will correctly update the cache.  This helps with many successive
writes.